apps/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "pydantic>=2.12.5",
    "ruff>=0.14.13",
    "uvicorn[standard]>=0.40.0",
    "openai>=1.60.0",
    "pytest>=8.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "algoliasearch>=4.35.4",
]

[dependency-groups]
dev = [
    "pytest-cov>=7.0.0",
]
